requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
certifi==2024.7.4
distro==1.9.0
Django==5.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
et-xmlfile==1.1.0
exceptiongroup==1.2.2
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
jiter==0.5.0
mysqlclient==2.2.4
numpy==2.0.1
openai==1.40.3
openpyxl==3.1.5
packaging==24.1
pandas==2.2.2
pydantic==2.8.2
pydantic_core==2.20.1
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
whitenoise==6.7.0
